requirements.txt
spotipy
requests
Pillow